Spit cast::string into a submodule of cast
